Fix for assertObjectEquals when the first object is a sparse array. Ensures that all properties are compared for both objects. Previously, if the first object was a sparse array, eg from an array literal like [,,a,,b,,] or an array object with partially full index space, then the missing indices would be ignored (and not compared) in both arrays.
Note: If you have a test that breaks because of this CL (despite our best due-diligence) then your test is faulty and should be fixed. You can verify this by swapping the order of the arguments to assertObjectEquelas, if it now fails (which it will) you have a faulty test.

R=pupius
DELTA=35  (19 added, 3 deleted, 13 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=208929


git-svn-id: http://closure-library.googlecode.com/svn/trunk@285 0b95b8e8-c90f-11de-9d4f-f947ee5921c8