This is another try
In the previous try, I was calling the static cleanup function
from an instance method of the test xhr object.  As a result,
the xhr object would be disposed sooner than it should have
been.

In this try, I changed it back so that cleanup is only called
when the READY event is fired, but now I ensure that the READY
event gets fired regardless of exceptions.

Original CL description:
This is a fix for a cleanup bug in the mock xhrio.

The problem occurs when the user throws an exception from his
xhr callback.  This is common when someone checks isSuccess()
and then throws an Error if isSuccess() is false.

In the mock version, the call to
simulateReadyStateChange(goog.net.XmlHttp.ReadyState.COMPLETE);
throws an exception, and as a result, simulateReady() is never called.
When simulateReady() isn't called, the cleanup() function is never
called, which leaves the static class in a bad state, causing future
tests to possibly fail and/or behave weirdly.

R=pallosp,nicksantos
DELTA=40  (25 added, 0 deleted, 15 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=806


git-svn-id: http://closure-library.googlecode.com/svn/trunk@757 0b95b8e8-c90f-11de-9d4f-f947ee5921c8