Automated g4 rollback
*** Reason for rollback ***

Suspect that it is causing regression

*** Original change description ***

Remove GC optimizations from goog.structs.Map.  This removes the parallel array of keys but in it's place introduces a map of numeric keys and an @define goog.structs.Map.PRESERVE_NON_STRING_KEYS that preserves the type information of the keys, for those who have come to depend on it.  Even though this goes against the documentation, changing it breaks a lot of tests.

R=pupius
DELTA=204  (62 added, 83 deleted, 59 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=3930


git-svn-id: http://closure-library.googlecode.com/svn/trunk@1454 0b95b8e8-c90f-11de-9d4f-f947ee5921c8