Fix maybeCreateDummyTargetForPosition_ to calculate clipping based on scrollableContainer_, if applicable.
This was causing problems for Music because we add the entire song list div as our target and then use the subtarget function to calculate the actual destination. So when maybeCreateDummyTargetForPosition_ cycled through the targets list to calculate clipping, it used the entire div box, which contained the mouse cursor, which caused it to be essentially ignored.

R=eae,pallosp,dgajda
DELTA=67  (66 added, 0 deleted, 1 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=3936


git-svn-id: http://closure-library.googlecode.com/svn/trunk@1460 0b95b8e8-c90f-11de-9d4f-f947ee5921c8