Changes some DateRange documentation to make it more clear.
As it is right now, the term "end date" is confusing. If I create a
DateRange that starts at midnight on January 1 and ends at midnight on
January 10, the 'endDate' parameter needs to be January 9, not January
10, which is confusing. I had to read the code to figure this out.

R=nnaze
DELTA=6  (0 added, 0 deleted, 6 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=4134


git-svn-id: http://closure-library.googlecode.com/svn/trunk@1541 0b95b8e8-c90f-11de-9d4f-f947ee5921c8